Adds changes from Qt Mobility Location repository.
There have been bug and documentation fixes applied to the Qt
Mobility repository since Qt Location was moved to the Qt 5
repository.

This commit is the result of a manual merge of those changes, with
care taken to try to avoid undoing any changes made when Qt Location
was moved into the Qt repo.