Add QGeoSearchManager unit test
This is from Qt Mobility as at a7f0eaff2295ced345fb05dafb31a2e6c5f98d6e

Note that the landmarkManager test case still fails because
we don't have the landmarkmanager plugins.  The plan is to make the
search manager into a geocoding manager and not need the landmarkmanager
anymore,  so later on we need to remove this test case.