remove dependency from landmarks from qgeoserachmanager(/engine)
Geocoding and searching for landmarks/places are separate concerns
and should not be combined together.  As a first step we remove
any dependency from landmarks for the searchmanager.
Later on we will need to refactor the searchmanager with another name
such as geocoding manager.

This change also makes the appropriate changes for the plugin
and tests