Refactoring to use QGeoLocation
This refactoring is comprised of 2 main tasks.
1) Use QGeoLocation instead of QPlaceLocation, which
came from the qt4-prerelease.
2) Use QGeoLocation instead of QGeoPlace from mobility.
QGeoPlace in mobility encapsulated an address + coordinate,
QGeoLocation in qt5 now fulfills that role(with some extra data fields)
Mostly the searchreply and searchmanagers have been refactored to
use qgeolocation instead of qgeoplace.

This change addresses the library, plugins and tests but only
some of the qml.

Note the geosearchmanager in future will need to be refactored into
a geocodingmanager, and there is still work needed on the qml
side of things.

Aside: the namings of mobility were used over the namings of the qt4
preprelease
eg QPlaceLocation::displayPosition->QGeoLocation::coordinate
   QPlaceLocatoin::mapView->QGeoLocation::viewport.

Change-Id: I312cddcb835ff52e94256c1fa97c6a9027eb9312
Reviewed-on: http://codereview.qt.nokia.com/1339
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>