Have QPlaceSearchQuery use a search area instead of coord or box
The original QPlaceSearchQuery accepts a coordinate or a boundingbox
to try narrow down the search.  When both coordinate and boundingbox
are set, it is ambiguous as to which one is used.  Also there
wasn't a way to specify a radius.

Instead we provide a single bounding area to eliminate any confusion.
Currently the bounding area can either be a circle or box.

Also add in a clone function for QGeoBoundingArea, this is so we can
pass clones into the QPlaceSearchQuery which takes ownership
of the search area.  The QPlaceSearchQuery takes ownership
so that queries can be easily created and destroyed on the stack,
(if the developer chooses do so, this is not mandatory)
otherwise developers would have to explicitly delete the search area.

This change does not address the issue of propagating notifications.
eg when a coord of a bounding box is changed, then
notifications that the area has changed should also be made.

Change-Id: I62656a6562d251e85ff967e96b2a4cbdb3cf7042
Reviewed-on: http://codereview.qt.nokia.com/1634
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>