Integrate changes from qt4-prerelease
This includes changes from c7c17f02b9cb1932dce60083627f01990e950ed5
to 291f466e02045f9b381bded49bda48cf35499297.

Mostly the changes are about updating the 'base' object for
the declarative elements when they are retrieved. i.e.  If a
the 'base' GeoLocation was retreived from the declarative geolocation,
then this is where fields such as the address and coordinate for
the base object are updated.  When the address for a declarative
geolocation is set, a declarative address is assigned but the base
geolocation is not updated at this point.

Also there's some changing of types from double to qreal for ratings
and scores.

Change-Id: Id8e16a977adf30a82927710c4f1e6e6b048f6d0c
Reviewed-on: http://codereview.qt.nokia.com/1814
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>