Add detailsFetched field to QGeoPlace
The details fetched field indicates whether the details of this
place have been retrieved or not.  An alternative may be to
return a timestamp of when the place has been retrieved,
but this may add extra complexity for no real benefit.

Also simplify the qmlplaces example so that only a single delegate
is used to how results.

Also change the default search location to Paris for easier
testing.  Le Marfil is populated with nearly all place data fields.

Change-Id: Ifc879258669c9f17b3002d6d4caf3c1b861eedbf
Reviewed-on: http://codereview.qt.nokia.com/1914
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>