Remove QPlaceContact and QGeoPlace::contacts functions
The primary purpose of QPlaceContact is to contain a contact
description and value.  The usefulness of this is when we
have multiple contact values and we need to distinguish between them

However the REST API does not support contact descriptions, so if
we retrieved  multiple phone numbers all we'd have is a set of
numbers and no way to tell the difference between them.
As a result returning a list of contacts is pointless as is the
QPlaceContact class.

So this change removes them.  A future change will add in functions
for primaryPhone(), primaryEmail() etc

Change-Id: I8b5bad35f1d8a0d23fe653d0bbed3a3fa366f5ea
Reviewed-on: http://codereview.qt.nokia.com/2233
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>