Share QGeoServiceProvider within QML.
QGeoServiceProvider is designed to provide cached instances of manager
objects, allowing a single manager to be used for the entire
application.  This feature was not being used for QML applications as
all of the declarative elements were creating a new QGeoServiceProvider
based on the data provided in the Plugin element.  This happened even if
the same Plugin element was used.

This commit adds a QGeoServiceProvider member to the Plugin element
which should be shared by all users of the Plugin element.

Change-Id: I508a14b8c9c8980baf1ffcb245cee13f6d25402d
Reviewed-on: http://codereview.qt.nokia.com/2244
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>