Refactor media to use map rather than a pagination class
The reason to use a map is so that we can store which media indexes
have been retrieved.  It is quite possible to retrieve non-contiguous
index, eg, a user might retrieve indexes 0-9, but then skips
a page and retrieves 10-19.  The key of the map is the index and the
value is the actual media object itself.

Change-Id: I0cc6aade3e38cbf8bbef707fb16c8f9b830660c0
Reviewed-on: http://codereview.qt.nokia.com/2404
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>