Remove BusinessInformation,BusinessFeature and weekdayhours
These have beens superseded by the PlaceAttribute class.

We do not have attribute keys for the closing notes,at they moment
they are pointless without a time period that is relevant for the
closing note.

We do not have attribute key for the opening hours yet since we have
yet to implement that in the plugin.

we do no yet attribute keys for business features because we never
had any to work with from the start.

Change-Id: I1a8749b1a598b4928a9c197cb5f4bde7615cdf8f
Reviewed-on: http://codereview.qt.nokia.com/2949
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>