Introduce Content class and rename PlaceMediaObject to PlaceImage
Paging is now implemented in terms of content objects, things like
images, reviews etc should inherit off the QPlaceContent class.
The main advantage of this change in C++ is that the paging code will
not be duplicated for different content types, also we have less
symbols which is good for plugin loading times.

For now, in QML we still retain separate specialized models.  In future
it might be possible to just have a single model class to hold
different content types.  There will still however need to be
separate model instances for say reviews and images.

PlaceMediaObject has been renamed to PlaceImage because currently
all we provide are images.  It is anticipated that if there are
new media types like video we would introduce QPlaceVideo and have
it inherit off QPlaceContent.  It is suboptimal to try define
a generic media object right now when we do not have a clear idea
about the use cases of each possible media type.

A future change will have QPlaceReview inherit off QPlaceContent.

Change-Id: Ia6cf5092e246374ed639694d7653e30429c94cc2
Reviewed-on: http://codereview.qt.nokia.com/3284
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>