Add initial jsondb plugin
Currently the plugin does not use proper location schema to store
place objects.  Also the plugin does some filtering in code rather
than with jsondb queries because it seems the comparison operators
are not working correctly.

These issues are to be addressed in future commits.  Currently
only basic place information is stored consisting of name, address
and coordinate.

Change-Id: Icc59f4fe50f5ecdd61d4d9565789d813a41a749f
Reviewed-on: http://codereview.qt.nokia.com/3555
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>