Rename PlaceSaveReply to PlaceIdReply
The id reply is meant for operation in which we return an id
such as saving a place, removing a category etc.

The id reply has an operationType() property which tells us what
the operation was.

We don't use the QPlaceReply::type() to hold the operation type
because the reply type should indicate the contents of the reply,
eg the content reply holds place content, the search reply
holds search results therefore the id reply holds an id.

Change-Id: I3aa0b45c648186ee78297620896420966e42afa3
Reviewed-on: http://codereview.qt.nokia.com/4061
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>