Add place add/updated/removed  signals + refactor Remove
These signals indicate changes to the database.  These 3 signals
are provided rather than a single placesChanged() signal because
it gives greater granularity.  E.g. if a place is removed,
we may want to know precisely which one it is so we can remove
it from a model.

For the SearchResultmodel whenever a place is added, updated, or
deleted we reload the model by executing the query again.

For now in the jsondb plugin implement signal emission
manually whenever a request is finished.  In a future change
we will use proper jsondb notifications.

For the online plugins, we do not expect these signals to be emitted.

Also place removal now uses the QPlaceIdReply.

Change-Id: I29f5aaaa6b8152abe1f5305a4d52abc721e6915e
Reviewed-on: http://codereview.qt.nokia.com/4109
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>