Added code to do the frustum/map intersection correctly.
This works around a problem with the QPolygonF
intersection code that we were having recently.

It also means that tilting a 2D map no longer
fetches all of the tiles to the edge of the map
within the field of view (ie clips tile fetching
at the far plane properly).

However, the tilting has exposed that there's a
bug in the code that works out which tiles to
fetch based on the bounding surface, so I'll
try to fix that next.