Rename QGeoPlace to QPlace
This is to be more consistent with the other place classes
like QPlaceManager.

The alternative would be for everything to have the QGeo prefix
e.g QGePlaceManager, QGeoPlaceCategory,
the Geo prefix does not help much in avoiding name collisions
since 'Place' shouldn't collide with anything anyway.

For classes like routing the Geo prefix is necessary to give
context for what's being routed.

Change-Id: Ibcf6e6909a41ea0ce2c402e24b427496d06bf243
Reviewed-on: http://codereview.qt-project.org/5210
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>