Remove additionalData from QPlace
additionalData is now superseded by place attributes.

In the qjsondetailsparser processAdContentPackages used to put
some content into additional data. indicating whether a place type
was an "ad_prime_plus" place.  For the Qt 5 use case, for 3rd party
developers this doesn't seem necessary and neither does it seem
part of the new REST API.

Also it seems that the search type has erroneously been placed into
the additionalData.  This is not necessary as the SearchResult
already has the type defined.

Change-Id: I35545b48666700a08aa93bfa0fa90ef2f301163e
Reviewed-on: http://codereview.qt-project.org/5213
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>