Add Category Save and Removal functionality
Note that the jsondb implementation will be heavily refactored.
The currrent implementation of having child id's within each
category does not 'gel' very well with jsondb.
(Some of the current engine functions are practically
synchronous)

Apparently there is an alternative approach using 'Views',
but further investigation is required.  For now the intention
is just to have something working so we can with experiment and
refine the API.

Change-Id: I5c9566e94f1a712f33dc06f6ad5e041326265398
Reviewed-on: http://codereview.qt-project.org/5215
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: abcd <qt_abcd1@ovi.com>