Allow search to take multiple categories
If multiple categories are provided it is expected that the backend
will effectively perform a union operation.  i.e. any places
that match one or more of the categories are returned.

It is possible however that some backends may not support
multiple categories.  In this case the first category is used and the
rest ignored.

Change-Id: I50133a1945a775d6961f0ada3b07ef1721d4de24
Reviewed-on: http://codereview.qt-project.org/6567
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>