Refactor contacts.
The REST server now supports labels for each contact value, so
now it makes sense to incorporte the ContactValue class.
ContactValue = user visible label + value(phone number, email etc)

We use a map type implementation where the key for the map
is the contact type and the value is a list of contact values.
The first element of each list becomes the primary value.
Eg the primaryPhone() field references the first element in the
list of phone numbers.

The setters for the primary___() functions has been removed.

Url has also been renamed to Website to keep in line with LBSP.

Change-Id: I9382a8286c5348bd1eb93d93f25088dfdf321852
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>