Docs for request classes + refactor
Documentation for the content and search request classes.

There is no need for the request classes to inherit off a
base request class.  Currently the base class has limit and offset
fields, but if for example there is an image submission request class
the limit and offset become meaningless.

The content reply class should have a field for the content request
just like the search reply has  a field for the search request.

Change-Id: I8d9dfe8a31806c6bb3b31f66d1338c2999e27ce2
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>