Rename Rating to Ratings
It may be possible that other rating values
could be added to the rating object e.g.
a count for how many gave a rating 1
a count for now many gave a rating of 2

or perhaps ratings for different aspects
of a place eg.
service
cleanliness
etc.

Therefore ratings should be used instead or rating.

Change-Id: I8a46646942dcd938bcee31921cabfa9df9862adc
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>