Fix place icon usage and supplier construction.
Place icons support either a full URL or a base URL from which an
icon URL is constructed.  Setting the full URL and base URL are
mutually exclusive, setting one resets the other.  Fix code which does
not take this into account and calls setBaseUrl() followed immediately
by setFullUrl().

Related is how the "nokia" plugin constructs QPlaceSupplier objects.
Fix two issues with this code.  First the base URL of the supplier icon
is set, even though only full URLs are supported.  Changed to set the
full URL.  Second, all suppliers are added to a supplier repository.
However, the repository did not behave as expected when a supplier was
already in the repository.  Ensure that missing data is copied into the
existing supplier stored in the repository.

Change-Id: I260d71f338e4cc0f5406b3a1351a215ada02b118
Reviewed-by: abcd <amos.choy@nokia.com>