Fix duplicate results
The issue was that declarative places were being cached
as a Map<QString, QDeclarativePlace *>.  However the search results
may be 'locations', i.e. places that consist of a name, address, coord
but have no other fields.  They didn't have a place id so when the
cached map was indexed, the same location was returned leading to
duplicates.  Instead cache the declarative places in a QList.

Change-Id: Id759d99a95ebb7388b8da03e9fdf299b7d6b5f99
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>