Split declarative tests into core and ui tests.
Sometimes it is desirable to run tests with the QT_QPA_PLATFORM=minimal
platform.  However, the minimal platform does not support OpenGL
contexts causing tests that depend on an OpenGL context to segfault.

This separates the tests into core and ui tests.  The ui tests can be
disabled by passing CONFIG+=no_ui_tests to qmake.

Change-Id: Ic756c8c572ae95aa403d2404c3a460cf7ab0d9eb
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>