Fixes the threading issues with maps.
The threading is about to be reworked so that multiple maps can
make use of the same mapping manager engine, which will change the
way this works quite a bit.

That means that this change is effectively a temporary fix to
unblock the CI, although aspects of this change will most likely
reappear in the new version.

Change-Id: I34377e69c15ac8de350308b99d84f8bd121e145c
Reviewed-by: Juha Vuolle <juha.vuolle@nokia.com>