No need  to rexecute query on every added/updated/removed place
The model does not need to be rexecuted time everytime a place changes.

For a update, we get the place to call getDetails
For a removal, we get the model to remove a place.

For an added place we don't do anything, we shouldn't be filtering
whether it is relevant to the search query or not, nor should
the query be re-executed.

Change-Id: Ie60c2d229c151459e55e33a4fe2f8eca251e77e6
Reviewed-by: abcd <amos.choy@nokia.com>