Unit tests + corrections for Contact Details in QML
The behaviour for contact details wasn't quite correct in the instance
where a single detail is assigned rather than a list.
For now there is a workaround by listening to the valueChanged signal
of the QDeclarativePropertyMap to detect if a single item
is being added and if so turn it into a list with one element.

The correction solution would be to override the write behaviour
of QDeclaratiePropertyMap but this can't be done until
QTBUG-23183 is addressed.

Also add some extra examples for contact detail assignment and
a correction for when contact details are viewed by a ListView,
we need to have modelData.label rather than just label when
dealing with an object list.

Change-Id: I10555b9536ae4c4172b9b49347b7b6be12c5c1ab
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>