Change Map#coordinateToScreenPosition to return closest solution
Coordinates can map to infinite screen positions if you allow
negative and very positive coordinates far beyond display. The first
solution (used previously) is not always the "best" for some situations,
particularly when map objects are close to the dateline. Choosing the
closest solution is not always perfect either (eg for huge objects),
but it is better and will handle a lot more common use cases.

Task-number: QTBUG-23368
Change-Id: Iab76fff07facc67cbce8cf20239d7ce413de8636
Reviewed-by: David Laing <david.laing@nokia.com>