Change Plugin{} element API to use enums for features
Notably, this change also sets up Map{} elements to delay probing their
set plugin until the plugin is complete, so that requiredFeatures
can function correctly without support for changing plugins
dynamically.

Please note: the necessary docs for these new methods and updates
to the already extant docs related to this will be in a separate
commit.

Change-Id: I36ae230ecdc07e3499331dddb251a331637405cd
Reviewed-by: David Laing <david.laing@nokia.com>