Refactoring ahead of precision issue fixes.
We need to rework the GL geometry and camera positioning code
to avoid problems with arising from the single precision
floats that GL uses.

This change is the first step down this path, which adds
double precision versions of the QVector2/3D classes and
changes the maps code to use them.

It also adds a new class which determines which tiles
will be visible based on the relevant bits of
information from the map (now using double precision).

This was done to force the decoupling of the code which
determines the visible tiles from the details of the
GL geometry / camera / etc...

Change-Id: I7127e39097a92028dfea01d816c14c9b9666569e
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>