Unify behaviour of BoundingBox and BoundingCircle elements.
QObject based properties of BoundingBox effectively had value sematics,
i.e. the object's properties were updated not the object pointer.  Both
BoundingBox and BoundingCircle now update the object pointers.

The canonical property values are stored in the QGeoBoundingBox and
QGeoBoundingCircle data members, these are then synchronised with the
QObject based properties.

A number of optimisations have also been made.  QObject based properties
are initialised on read, meaning that if a particular property is not
used it will not be constructed.  Value properties (width, height,
radius) are cached so that repeated property reads do not result in
multiple nested function calls.

It is also possible to make a valid bounding box by setting its center
coordinate.  The resulting bounding box will have a width and height
of 0 degrees.

Change-Id: I2569452e9a92c0b84de1b8f690655e2bed19ecb1
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>