Support more than QObject* based roles in MapItemView.
This change makes the MapItemView handle models like the views in
QtQuick.  Note that only QAbstractItemModel based models still the
only supported models.

Adds support for model roles with data types other than QObject* ones.
The model roles are exposed as context properties in the delegate or
via the "model" context property, which is created to mitigate against
name clashes.

Change-Id: I5527609e2ae58be11fcffb93060be62e8a66ea7e
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>