Correctly implement initialize categories in jsondb plugin
For ease of use the category functions are synchronous,
to achieve this we keep a cached version of the categories
when initializeCategories is called.

We don't expect there to be a lot of categories and
so there shouldn't be much memory consumption.

As a result of these cached categories we no longer
need to use an event loop whenever categories
are queried.  The event loop was problematic when
multiple calls were made to the function while
the event loop was still active.

Change-Id: Iff8482ce9affcec82bee5c12b325e64b77624c65
Reviewed-by: abcd <amos.choy@nokia.com>