Fix for QTBUG-23985
Uses the pixel offset of map objects to make sure that the correct
behaviour occurs with mouse clicks on the top and left edges of
objects such as polygons, rectangles and circles.

This is rather annoying to do correctly, as the triangulators for
the fill and border each do their own clipping separately and may
not return the same clipping rect.

Task-number: QTBUG-23985
Change-Id: I862ef439ec3ed4789caf363de1f1778887fb93df
Reviewed-by: David Laing <david.laing@nokia.com>