Regenerate map object geometry on Map resize event
In some cases (eg static map objects in the Mapviewer example),
the resize event can race against the map setup routines and lose,
leading to lots of NaNs, invisible map objects or missing borders.

To remedy this as a temporary fix, we simply check for whether the
map size has changed in handleCameraDataChanged. This API probably
needs to be rethought in the near future.

Change-Id: I67a2132c756bb38bf1513fa1cf3efc3ccb5de0da
Reviewed-by: David Laing <david.laing@nokia.com>