Searching by bounding box provides weird results. QTBUG-23735
There was a bug in the json code with the distance calcualtion. The
distance was not being calculated at all. Also removed a flag that was
used to copy the results into the list as this was unecessary.

This is a temporary fix. The entire file needs to be cleaned up.

Change-Id: I58f53fac2611749bafbe075a40afb9ce7ef8285b
Reviewed-by: Jean Gressmann <jean.gressmann@nokia.com>
Reviewed-by: abcd <amos.choy@nokia.com>