Convert QGeoBoundingArea to use value type semantics.
This simplifies the API while keeping the same functionality.
Objects of type QGeoBoundingCircle and QGeoBoundingBox can still
be passed around as QGeoBoundingArea without losing their identity.
QGeoBoundingArea::isValid(), ::isEmpty() and ::contains() function
as one would expect.

This change removes the necessity of the application developer to
track and delete pointers.

Change-Id: I1ee70af90319da4c77225c6a0a03601d5ff431f1
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>