Fix compiler errors when QT_PLACES_LOGGING is defined.
It woud appear that the debugging QT_PLACES_LOGGING macro, when added
and enabled by definition when added to the pro file would cause
compilation errors. This is perhaps old code, but it would nonetheless
be nice to have the code working and updated for the current version
with the changes made to the base code.

Change-Id: I97048bbb14ee5b006e89b3a98c6b8c27085cd143
Reviewed-by: Alex <alex.blasche@nokia.com>