Remove the historical referer from the mapping service.
The maptile service previously required a token, and optionally a
referer to be set in the request. The new authentication system requires
a token and app_id. Leaving the referer in the request causes confusion,
especially now the the token has a new meaning and is also validated
differently.

This patch will remove the refererer from the mapping service only.

Change-Id: I2aff52ea74692af05cf381a48376b13ec1ba4bb7
Reviewed-by: Cristian Adam <cristian.adam@nokia.com>
Reviewed-by: Alex <alex.blasche@nokia.com>