Fix up issues with parsing Route XML in Nokia plugin
This fixes some long-standing issues with route XML parsing in the
Nokia plugin, as well as adding a new autotest to verify that these
and other faults do not develop again in future.

The faulty logic in this case was to attempt to handle the final
maneuver in the route in the same manner as a link-less intermediate
point, which alternately caused the final maneuver to either be
dropped or a subsequent "empty" maneuver to be generated with the
remaining links at the end. The final maneuver is now handled
separately.

Task-number: QTBUG-24341
Task-number: QTBUG-20563
Change-Id: Iefc37c84170d23b4c4a9203f12f33994c7fa4a88
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>