Fix for pipelined tile requests receiving incorrect data
Calling abort() and then deleteLater() without waiting for finished()
and friends to come back turns out to be a Bad Idea. QNAM won't be
able to correctly handle the abortion if the request is occupying a
slot in a pipeline queue, and that queue will be off-by-one for the
remainder of its operation.

Fix is to change the finished() connection to be Queued, and simply
call abort rather than deleteLater() as well. Then in the queued
finished() handler, deleteLater() will be called and all will be
well (hopefully).

Change-Id: Ia572757af8394297fa707f0d8d1accded2eec12d
Reviewed-by: David Laing <david.laing@nokia.com>