Reset view index when PlaceDialog is closed
Previously if a user switched to the categories page and clicked
close the dialog stayed on that page.

So now reset to the initial page whenever the opacity reaches 0.
An immediate switch to the initial page is not done because the
transition is viewable as the close button is pressed, so wait
until the dialog has faded out of view before switching.

Change-Id: I1b96fc95a309edbc1a739e4ceddb38a6b25b6cba
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>