Document PlaceSearchModel behavior when it detects update/removal
The PlaceSearchModel will get the details of any place that it detects
an update on and automatically remove any place that has been removed.

A by product of this is that whenever a place from a model is saved,
it is always followed by a fetch, leading to a sequence of state
changes: Saving, Ready, Fetching, Ready.  This quirk needs to
be clearly documented in addition to the PlaceSearchModel behavior.

Change-Id: If15a3a4250a45a9ba51b815233f21fefcbc1ee47
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>