Remove use of Loader element in Dialogs.
Simplify the code by removing the Loader element.  There is really no
need for it as a content item can be constructed directly.

Factor out the code for text field entry dialogs into a separate
InputDialog element.  Dialogs that create a custom content item no
longer carry around unneeded API.

Account for the possibility of an on screen virtual keyboard.

Change-Id: I808e040c264b8c8389a03070cc628e2e4518d118
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>
Reviewed-by: abcd <amos.choy@nokia.com>