Fix assigning a category to a place in example
A snap mode doesn't need to be set for the view since it
is not interactive.

Assignment of a category caused the view to resize itself.
This resizing affected the snap mode logic, resulting
in a situation where the view did not follow the currentIndex,
see QTBUG-25176.  The symptom was that once a category was assigned
the currentIndex of the view could no longer be changed.

Change-Id: I036cccd38c3059bb357c02715b188ab4408a6cc7
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>