Make QGeoCoordinate a QSharedData class
We deal with a lot of QGeoCoordinates on the stack as value types,
and their d_ptr is always allocated on the heap. This creates a lot
of unnecessary heap churn, which QSharedData can avoid. Also, with
this patch applied, the heap usage component of QGeoCoordinate on
the MapViewer example is reduced from ~1.4Mb to ~600kb.

Change-Id: I6675e861ca937667e4b66dfae6950ba4de22bd80
Reviewed-by: Ian Chen <ian.1.chen@nokia.com>
Reviewed-by: Thomas Lowe <thomas.lowe@nokia.com>
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>