'Map panning sometimes looks like flick action' fixed
Firstly the time was reduced from 100ms to 50ms at which to consider the
gesture a flick, so a pan is less likely to be mistaken for a flick.
Secondly, the current method of storing the last 3 velocities was flawed as
the updates can be so frequent that the elapsed time can be 0 or 1 millisecond,
the limit of timer accuracy. This caused bad estimations of the velocity.
This has been replaced with a single sample that updates every 50ms.

Change-Id: I5c755e2978e7730475043187de0440cc411c934f
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>