Fix crash when search suggestion is canceled.
There seems to have been some sort of race condition
where m_reply is not valid, but is still being operated on
when cancel is called.

To avoid this always set m_reply to 0 when it has been
canceled.

Note, the abort operation can potentially (albeit indirectly)
set m_reply to 0 which is why there is a check
after QPlaceReply::abort() is called.

Change-Id: I46aa5b4eb6cbf2acba1a57f63ec073f56bd0f386
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>