New triangulator for map polygons
Replacing the qTriangulate call with poly2tri. This is a BSD-licensed
triangulation library that gives better performance for most of our
shapes. See http://pastie.org/3893553 for a timing comparison on
linux-x86 (timing is given for the whole screen transformation step)

The BSD-licensed code is isolated into its own dir (in src/3rdparty)
and compiled as a separate static library to be linked in. The files
have been renamed (.cc => .cpp) and doubles changed to floats, but
are otherwise identical to those found at http://code.google.com/p/poly2tri/

Change-Id: Ib07aadafdc5d4b8ac1024e1eecd70ab39f673197
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>