Refactor to allow specification for nokia icons only
There is an plugin parmeter called places.icons.custom
that defines whether custom 3rd party icons or nokia
icons are presented to the user for the jsondb plugin.

For the nokia plugins, nokia icons can be subdivded
into two types. The first is explictly set icons, the second
is icons generated from the categories.

In both cases the base icon is identified by the
nokiaIcon key and a base icon value.
However in the case of a generated icon there is
a nokiaIconGenerated key that is set to true.

When saving a place to the jsondb plugin, if
the icon is generated, the icon data is not saved to jsondb.
The category data however always is.  When the place is
retrieved from jsondb, the base icon will be generated from the
category data.

If an icon is not generated, then the base icon is saved
to jsondb under the "nokiaIcon" field.  When
the place is retried from jsondb, this field will be used
to provide the base icon.

Change-Id: I9100011c519185719cdd7b62e12c2386f7f0f8d7
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>