Fix case where second and subsequent searches fail.
When a request is cancelled or fails the emit finished is called. This
causes any subsequent requests to complete immediately, resulting in no
more searches working. Also, the initial state of the plugin should be
Null which correctly matches the documentation.

The state change after the cancel should be overwritten to cancel.

Change-Id: I22c2ffe92fbe70e6391ae4f049368edaeafb1c4d
Reviewed-by: Jean Gressmann <jean.gressmann@nokia.com>
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>