Save and restore cache queues, for better caching
The particular cache queues that each tile was in are now restored on startup
this is a more effective use of the cache.

It relies on the mappingManagerEngine's destructor getting called on closing.
This doesn't currently happen and is a separate Jira bug. However, the cache will
still load up, so nothing is broken in the meantime.

Task-number: QTBUG-25135
Change-Id: I438fc9980506f9315b6f1934589c05563c242c15
Reviewed-by: Alex Wilson <alex.wilson@nokia.com>
Reviewed-by: Ian Chen <ian.1.chen@nokia.com>